0x1E-search_algorithms

+++ 0x1E-search_algorithms/0-linear.c
@@ -11,12 +11,12 @@
 
 int linear_search(int *array, size_t size, int value)
 {
-	size_t i = 0;
+	int i = 0;
 
 	if (!array)
 		return (-1);
 
-	for (; i < size; i++)
+	for (; i < (int) size; i++)
 	{
 		printf("Value checked array[%d] = [%d]\n", i, array[i]);
 		if (value == array[i])
